fix(planner): keep chat_service under 1000 and accept empty plan_lines
Move the session today-line field onto ChatTodayLine so the god-file ratchet stays closed. Card goldens now include the new empty plan_lines key.
