+++ lib/services/chat/chat.dart
@@ -86,6 +86,7 @@ export 'stop_sequences.dart';
 export 'story_clock.dart';
 export 'story_clock_claims.dart';
 export 'time_service.dart';
+export 'chat_today_line.dart';
 export 'today_line_tag.dart';
 export 'tool_support_tester.dart';
 export 'weather_biomes.dart';
