+++ test/golden/_goldens/card/v25_full.golden.json
@@ -46,7 +46,8 @@
         "needs_sim_strength": 1,
         "nsfw_cooldown_enabled": false,
         "passage_of_time_enabled": true,
-        "realism_needs_director_authority": false,
+                "plan_lines": [],
+"realism_needs_director_authority": false,
         "realism_verification_enabled": false,
         "realism_verification_max_reprocesses": 1,
         "realism_verification_strictness": 3,
